MTDSA-23692: updated code and ITs

+++ app/v5/retrieveCalculation/def2/model/response/calculation/transitionProfit/TransitionProfit.scala
@@ -19,7 +19,7 @@ package v5.retrieveCalculation.def2.model.response.calculation.transitionProfit
 import play.api.libs.json.{Format, Json}
 
 case class TransitionProfit(
-    totalTaxableTransitionProfit: Option[BigDecimal],
+    totalTaxableTransitionProfit: Option[Long],
     transitionProfitDetail: Option[Seq[TransitionProfitDetail]]
 )
